changed a couple of files, deleted some
